+++ CHANGELOG.md
@@ -8,6 +8,16 @@
 * Skip confirming patches that contain only a datamodel name change. ([#688])
 * Added sync reminder notification. ([#689])
 * Added protection against syncing a model to a place. ([#691])
+* Select Instances from diff tree view ([#709])
+* Fix Rojo breaking when users undo/redo in Studio ([#708])
+* Improved sync info text on Connected page. ([#692])
+* Fix patch visualizer breaking when instances are removed during sync ([#713])
+* Patch visualizer now indicates what changes failed to apply. ([#717])
+* Add buttons for navigation on the Connected page ([#722])
+* Improve tooltip behavior ([#723])
+* Better settings controls ([#725])
+* Rework patch visualizer with many fixes and improvements ([#726])
+* Added support for syncing in `.toml` files ([#633])
 
 [#631]: https://github.com/rojo-rbx/rojo/pull/631
 [#668]: https://github.com/rojo-rbx/rojo/pull/668
@@ -16,6 +26,16 @@
 [#688]: https://github.com/rojo-rbx/rojo/pull/688
 [#689]: https://github.com/rojo-rbx/rojo/pull/689
 [#691]: https://github.com/rojo-rbx/rojo/pull/691
+[#709]: https://github.com/rojo-rbx/rojo/pull/709
+[#708]: https://github.com/rojo-rbx/rojo/pull/708
+[#692]: https://github.com/rojo-rbx/rojo/pull/692
+[#713]: https://github.com/rojo-rbx/rojo/pull/713
+[#717]: https://github.com/rojo-rbx/rojo/pull/717
+[#722]: https://github.com/rojo-rbx/rojo/pull/722
+[#723]: https://github.com/rojo-rbx/rojo/pull/723
+[#725]: https://github.com/rojo-rbx/rojo/pull/725
+[#726]: https://github.com/rojo-rbx/rojo/pull/726
+[#633]: https://github.com/rojo-rbx/rojo/pull/633
 
 ## [7.3.0] - April 22, 2023
 * Added `$attributes` to project format. ([#574])